project_root/requirements.txt
streamlit
boto3
pywhatkit
twilio
Pillow
psutil
requests
beautifulsoup4
selenium
webdriver-manager
tweepy
python-dotenv
pandas
instagrapi
